refactor(frontend): extract logic into custom hooks and split large UI components
